Coordinates class moved to core.py. Coordinates implemented in most of Equilibrium functions !still not to all!
