--- pleque/coordinates.py
@@ -1,235 +1,0 @@
-import numpy as np
-
-from pleque.core import Equilibrium
-
-
-class Coordinates(object):
-
-    def __init__(self, equilibrium: Equilibrium, *coordinates, coord_type=None, grid=False, **coords):
-        self._eq = equilibrium
-        self._valid_coordinates = {'R', 'Z', 'psi_n', 'psi', 'rho'}
-        self._valid_coordinates_1d = {('psi_n',), ('psi',), ('rho',)}
-        self._valid_coordinates_2d = {('R', 'Z')}
-        self.dim = -1  # init only
-        self.grid = grid
-
-        self.__evaluate_input__(*coordinates, coord_type=coord_type, **coords)
-
-    def __call__(self, *args, **kwargs):
-        pass
-
-    def __iter__(self):
-        pass
-
-    def sort(self, order):
-        pass
-
-    @property
-    def R(self):
-        if self.dim == 2:
-            return self.x1
-
-    @property
-    def Z(self):
-        if self.dim == 2:
-            # todo
-            return self.x2
-
-    @property
-    def psi(self):
-        if self.dim == 1:
-            return self._eq._psi_axis + self.x1 * (self._eq._psi_lcfs - self._eq._psi_axis)
-        elif self.dim == 2:
-            self._eq._spl(self.x1, self.x2, grid=self.grid)
-
-    @property
-    def psi_n(self):
-        if self.dim == 1:
-            return self.x1
-        elif self.dim == 2:
-            return (self.psi - self._eq._psi_axis) / (self._eq._psi_lcfs - self._eq._psi_axis)
-
-    @property
-    def rho(self):
-        return np.sqrt(self.psi_n)
-
-    # todo
-    # @property
-    # def r_mid(self):
-    #     """
-    #     Midplane coordinate.
-    #     :return:
-    #     """
-    #
-    #
-    #     return
-
-    def getAs(self, coord_type=None):
-        if self.dim == 0:
-            return np.array(())
-        coord_type_ = self._verify_coord_type(coord_type)
-
-        if self.dim < len(coord_type):
-            raise ValueError('Can not return 2d from 1d data.')
-        raise ValueError('Not implemented yet.')
-
-    def __evaluate_input__(self, *coordinates, coord_type=None, **coords):
-        from collections import Iterable
-
-        if len(coordinates) == 0:
-            # todo:
-            self.dim = 0
-            xy = []
-            xy_name = []
-            for key, val in coords.items():
-                if key in self._valid_coordinates and val is not None:
-                    if isinstance(val, Iterable):
-                        if not isinstance(val, np.ndarray):
-                            val = np.array(val, ndmin=1)
-                        if len(val.shape) == 0:
-                            val = val.reshape((len(val), 1))
-                    else:
-                        val = np.array(val, ndmin=1)
-                    xy.append(val)
-                    xy_name.append(key)
-                    self.dim += 1
-
-            if self.dim == 0:
-                coord_type_ = ()
-            elif self.dim == 1:
-                self._x1_input = xy[0]
-                coord_type_ = tuple(xy_name)
-            elif self.dim == 2:
-                if tuple(xy_name) in self._valid_coordinates_2d:
-                    self._x1_input = xy[0]
-                    self._x2_input = xy[1]
-                    coord_type_ = tuple(xy_name)
-                elif tuple(xy_name[::-1]) in self._valid_coordinates_2d:
-                    self._x1_input = xy[1]
-                    self._x2_input = xy[0]
-                    coord_type_ = tuple(xy_name[::-1])
-                    if len(self._x1_input) != len(self._x2_input):
-                        raise ValueError('All coordinates should contain same dimension.')
-                else:
-                    raise ValueError('Invalid combination of input coordinates.')
-
-            else:
-                # self._incompatible_dimension_error(self.dim)
-                raise ValueError('Operation in {} space has not be en allowed yet. Sorry.'
-                                 .format(self.dim))
-
-            self._coord_type_input = coord_type_
-        else:
-            if len(coordinates) == 1:
-                xy = coordinates[0]
-
-                if self.grid:
-                    print('WARNING: grid == True is not allowed for this coordinates input. '
-                          'Turning grid = False.')
-                self.grid = False
-
-                if isinstance(xy, Coordinates):
-                    # todo: ask Ondra (!)
-                    print('ERROR: not implemented yet')
-                    raise ValueError('not implemented yet')
-                    pass
-                elif isinstance(xy, Iterable):
-                    # input as array of size (N, dim),
-                    # if (N) add dimension
-                    if not isinstance(xy, np.ndarray):
-                        xy = np.array(xy, ndmin=2)
-                    if len(xy.shape) == 1:
-                        xy = xy.reshape((len(xy), 1))
-
-                    self.dim = xy.shape[1]
-                    if self.dim == 1:
-                        self._x1_input = xy[:, 0]
-                    elif self.dim == 2:
-                        self._x1_input = xy[:, 0]
-                        self._x2_input = xy[:, 1]
-                    else:
-                        self._incompatible_dimension_error(self.dim)
-                else:
-                    # 1d, one number
-                    self.dim = 1
-                    self._x1_input = np.array([xy])
-                pass
-            elif len(coordinates) == 2:
-                self.dim = 2
-                x1 = coordinates[0]
-                x2 = coordinates[1]
-
-                # assume _x1_input and _x2_input to be arrays of size (N)
-                if not isinstance(x1, np.ndarray):
-                    x1 = np.array(x1, ndmin=1)
-                if not isinstance(x2, np.ndarray):
-                    x2 = np.array(x2, ndmin=1)
-                self._x1_input = x1
-                self._x2_input = x2
-
-            else:
-                self._incompatible_dimension_error(len(coordinates))
-
-            self._coord_type_input = self._verify_coord_type(coord_type)
-        self._convert_to_default_coord_type()
-
-    def _verify_coord_type(self, coord_type):
-        if isinstance(coord_type, str):
-            coord_type = (coord_type,)
-
-        if self.dim == 0:
-            ret_coord_type = ()
-        elif self.dim == 1:
-            if coord_type is None:
-                ret_coord_type = ('psi_n',)
-            elif tuple(coord_type) in self._valid_coordinates_1d:
-                ret_coord_type = tuple(coord_type)
-            else:
-                ret_coord_type = ('psi_n',)
-                print("WARNING: _coord_type_input is not correct. \n"
-                      "{} is not allowed \n"
-                      "Force set _coord_type_input = ('psi_n',)"
-                      .format(tuple(coord_type)))
-        elif self.dim == 2:
-            if coord_type is None:
-                ret_coord_type = ('R', 'Z')
-            elif tuple(coord_type) in self._valid_coordinates_2d:
-                ret_coord_type = tuple(coord_type)
-            elif tuple(coord_type[::-1]) in self._valid_coordinates_2d:
-                ret_coord_type = tuple(coord_type[::-1])
-            else:
-                ret_coord_type = ('R', 'Z')
-                print("WARNING: _coord_type_input is not correct. \n"
-                      "{} is not allowed \n"
-                      "Force set _coord_type_input = ('R', 'Z')"
-                      .format(tuple(coord_type)))
-        else:
-            raise ValueError('Operation in {} space has not be en allowed yet. Sorry.'
-                             .format(self.dim))
-        # todo: make order dependent!
-        return ret_coord_type
-
-    def _incompatible_dimension_error(self, dim):
-        raise ValueError('Operation in {} space has not be en allowed yet. Sorry.'
-                         .format(dim))
-
-    def _convert_to_default_coord_type(self):
-        if self.dim == 0:
-            return
-        elif self.dim == 1:
-            # convert to psi_n
-            if self._coord_type_input == ('psi_n',):
-                self.x1 = self._x1_input
-            elif self._coord_type_input == ('psi',):
-                psi = self._x1_input
-                self.x1 = (psi - self._eq._psi_axis) / \
-                          (self._eq._psi_lcfs - self._eq._psi_axis)
-
-            elif self._coord_type_input == ('rho',):
-                self.x1 = self._x1_input ** 2
-            else:
-                raise ValueError('This should not happen.')
-        elif self.dim == 2:
-            # only (R, Z) coordinates are implemented now
-            self.x1 = self._x1_input
-            self.x2 = self._x2_input
