+++ pleque/core.py
@@ -97,7 +97,7 @@ class Equilibrium(object):
         # todo: first assume r, z are ascending:
         spl = RectBivariateSpline(r, z, psi, kx=spline_order, ky=spline_order,
                                   s=spline_smooth)
-        self._spl = spl
+        self._spl_psi = spl
 
         # find extremes:
         self.__find_extremes__()
@@ -133,56 +133,48 @@ class Equilibrium(object):
         :param coords:
         :return:
         """
-        if psi_n is not None:
-            return self._psi_axis + psi_n * (self._psi_lcfs - self._psi_axis)
-
-        return self._spl(R, Z, grid=grid)
+        coord = Coordinates(self, *coordinates, R=R, Z=Z, psi_n=psi_n, coord_type=coord_type, grid=grid, **coords)
+        return coord.psi
 
     def psi_n(self, *coordinates, R=None, Z=None, psi=None, coord_type=None, grid=True, **coords):
-        if psi is None:
-            psi = self.psi(R=R, Z=Z, grid=grid)
-        return (psi - self._psi_axis) / (self._psi_lcfs - self._psi_axis)
+        coord = Coordinates(self, *coordinates, R=R, Z=Z, psi=psi, coord_type=coord_type, grid=grid, **coords)
+        return coord.psi_n
 
     @property
     def _diff_psi_n(self):
         return 1 / (self._psi_lcfs - self._psi_axis)
 
-    def rho(self, R=None, Z=None, psi_n=None, coord_type=None, grid=True, **coords):
-        if R is not None and Z is not None:
-            psi_n = self.psi_n(R=R, Z=Z, grid=grid)
-        return np.sqrt(psi_n)
+    def rho(self, *coordinates, R=None, Z=None, psi_n=None, coord_type=None, grid=True, **coords):
+        coord = Coordinates(self, *coordinates, R=R, Z=Z, psi_n=psi_n, coord_type=coord_type, grid=grid, **coords)
+        return np.sqrt(coord.psi_n)
 
     def pressure(self, *coordinates, R=None, Z=None, psi_n=None, coord_type=None, grid=True, **coords):
-        if R is not None and Z is not None:
-            psi_n = self.psi_n(R=R, Z=Z, grid=grid)
-        return self._pressure_spl(psi_n)
+        coord = Coordinates(self, *coordinates, R=R, Z=Z, psi_n=psi_n, coord_type=coord_type, grid=grid, **coords)
+        return self._pressure_spl(coord.psi_n)
 
     def pprime(self, *coordinates, R=None, Z=None, psi_n=None, coord_type=None, grid=True, **coords):
-        if R is not None and Z is not None:
-            psi_n = self.psi_n(R=R, Z=Z, grid=grid)
-        return self._dp_dpsin_spl(psi_n) * self._diff_psi_n
+        coord = Coordinates(self, *coordinates, R=R, Z=Z, psi_n=psi_n, coord_type=coord_type, grid=grid, **coords)
+        return self._dp_dpsin_spl(coord.psi_n) * self._diff_psi_n
 
     def fpol(self, *coordinates, R=None, Z=None, psi_n=None, coord_type=None, grid=True, **coords):
-        if R is not None and Z is not None:
-            psi_n = self.psi_n(R=R, Z=Z, grid=grid)
-        mask_out = psi_n > 1
-        fpol = self._fpol_spl(psi_n)
+        coord = Coordinates(self, *coordinates, R=R, Z=Z, psi_n=psi_n, coord_type=coord_type, grid=grid, **coords)
+        # todo use in_plasma
+        mask_out = coord.psi_n > 1
+        fpol = self._fpol_spl(coord.psi_n)
         fpol[mask_out] = self.BvacR
         return fpol
 
     def fprime(self, *coordinates, R=None, Z=None, psi_n=None, coord_type=None, grid=True, **coords):
-        if R is not None and Z is not None:
-            psi_n = self.psi_n(R=R, Z=Z, grid=grid)
-        mask_out = psi_n > 1
-        fprime = self._df_dpsin_spl(psi_n) * self._diff_psi_n
+        coord = Coordinates(self, *coordinates, R=R, Z=Z, psi_n=psi_n, coord_type=coord_type, grid=grid, **coords)
+        mask_out = coord.psi_n > 1
+        fprime = self._df_dpsin_spl(coord.psi_n) * self._diff_psi_n
         fprime[mask_out] = 0
         return fprime
 
     def ffprime(self, *coordinates, R=None, Z=None, psi_n=None, coord_type=None, grid=True, **coords):
-        if R is not None and Z is not None:
-            psi_n = self.psi_n(R=R, Z=Z, grid=grid)
-        mask_out = psi_n > 1
-        ffprime = self._fpol_spl(psi_n) * self._df_dpsin_spl(psi_n) * self._diff_psi_n
+        coord = Coordinates(self, *coordinates, R=R, Z=Z, psi_n=psi_n, coord_type=coord_type, grid=grid, **coords)
+        mask_out = coord.psi_n > 1
+        ffprime = self._fpol_spl(coord.psi_n) * self._df_dpsin_spl(coord.psi_n) * self._diff_psi_n
         ffprime[mask_out] = 0
         return ffprime
 
@@ -209,15 +201,16 @@ class Equilibrium(object):
         :param coords:
         :return: Absolute value of magnetic field in Tesla.
         """
-        B_R = self.B_R(R=R, Z=Z)
-        B_Z = self.B_Z(R=R, Z=Z)
-        B_T = self.B_tor(R=R, Z=Z)
+        coord = Coordinates(self, *coordinates, R=R, Z=Z, coord_type=coord_type, grid=grid, **coords)
+        B_R = self.B_R(R=coord.R, Z=coord.Z)
+        B_Z = self.B_Z(R=coord.R, Z=coord.Z)
+        B_T = self.B_tor(R=coord.R, Z=coord.Z)
         B_abs = np.sqrt(B_R ** 2 + B_Z ** 2 + B_T ** 2)
 
         return B_abs
 
-
-    def B_R(self, *coordinates, R=None, Z=None, coord_type=None, grid=True, **coords):
+    # todo: resolve the grids
+    def B_R(self, *coordinates, R=None, Z=None, coord_type=('R', 'Z'), grid=True, **coords):
         """
         Poloidal value of magnetic field in Tesla.
 
@@ -228,12 +221,13 @@ class Equilibrium(object):
         :param coords:
         :return:
         """
+        coord = Coordinates(self, *coordinates, R=R, Z=Z, coord_type=coord_type, grid=grid, **coords)
         if grid:
             Rs, Zs = np.meshgrid(R, Z)
             Rs = Rs.T
         else:
             Rs = R
-        return -self._spl(R, Z, dy=1, grid=grid) / Rs * self._Bpol_sign
+        return -self._spl_psi(R, Z, dy=1, grid=grid) / Rs * self._Bpol_sign
 
     def B_Z(self, *coordinates, R=None, Z=None, coord_type=None, grid=True, **coords):
         """
@@ -246,13 +240,14 @@ class Equilibrium(object):
         :param coords:
         :return:
         """
+        coord = Coordinates(self, *coordinates, R=R, Z=Z, coord_type=coord_type, grid=grid, **coords)
         if grid:
             Rs, Zs = np.meshgrid(R, Z)
             Rs = Rs.T
         else:
             Rs = R
 
-        return self._spl(R, Z, dx=1, grid=grid) / Rs * self._Bpol_sign
+        return self._spl_psi(R, Z, dx=1, grid=grid) / Rs * self._Bpol_sign
 
     def B_pol(self, *coordinates, R=None, Z=None, coord_type=None, grid=True, **coords):
         """
@@ -265,6 +260,7 @@ class Equilibrium(object):
         :param coords:
         :return:
         """
+        coord = Coordinates(self, *coordinates, R=R, Z=Z, coord_type=coord_type, grid=grid, **coords)
         B_R = self.B_R(R=R, Z=Z)
         B_Z = self.B_Z(R=R, Z=Z)
         B_pol = np.sqrt(B_R ** 2 + B_Z ** 2)
@@ -281,6 +277,7 @@ class Equilibrium(object):
         :param coords:
         :return:
         """
+        coord = Coordinates(self, *coordinates, R=R, Z=Z, coord_type=coord_type, grid=grid, **coords)
         if grid:
             R_mesh = R[:, None]
         else:
@@ -334,11 +331,11 @@ class Equilibrium(object):
         rs = np.linspace(self.r_min, self.r_max, 120)
         zs = np.linspace(self.z_min, self.z_max, 130)
 
-        psi = self.psi(R=rs, Z=zs)
-        psi_x = self._spl(rs, zs, dx=1, dy=0)
-        psi_y = self._spl(rs, zs, dx=0, dy=1)
+        psi = self._spl_psi(rs, zs)
+        psi_x = self._spl_psi(rs, zs, dx=1, dy=0)
+        psi_y = self._spl_psi(rs, zs, dx=0, dy=1)
         psi_xysq = psi_x ** 2 + psi_y ** 2
-        psi_xy = self._spl(rs, zs, dx=1, dy=1)
+        psi_xy = self._spl_psi(rs, zs, dx=1, dy=1)
 
         mins0 = tuple(argrelmin(psi_xysq, axis=0))
         mins1 = tuple(argrelmin(psi_xysq, axis=1))
@@ -351,8 +348,8 @@ class Equilibrium(object):
         # plt.show()
 
         def psi_xysq_func(x):
-            return self._spl(x[0], x[1], dx=1, dy=0, grid=False) ** 2 \
-                   + self._spl(x[0], x[1], dx=0, dy=1, grid=False) ** 2
+            return self._spl_psi(x[0], x[1], dx=1, dy=0, grid=False) ** 2 \
+                   + self._spl_psi(x[0], x[1], dx=0, dy=1, grid=False) ** 2
 
         x_points = []
         o_points = []
@@ -375,7 +372,7 @@ class Equilibrium(object):
                     z_ex2 = res['x'][1]
 
                     #                    psi_xyabs = np.abs(psi_xy[ar, az])
-                    psi_xyopt = np.abs(self._spl(r_ex2, z_ex2, dx=1, dy=1, grid=False)) ** 2
+                    psi_xyopt = np.abs(self._spl_psi(r_ex2, z_ex2, dx=1, dy=1, grid=False)) ** 2
 
                     if psi_xyopt < 0.1:
                         # plt.plot(rs[ar], zs[az], 'o', markersize=10, color='b')
@@ -396,13 +393,13 @@ class Equilibrium(object):
 
         op_dist = (o_points[:, 0] - r_centr) ** 2 + (o_points[:, 1] - z_centr) ** 2
         # assume that psi value has its minimum in the center
-        op_psiscale = self.psi(R=o_points[:, 0], Z=o_points[:, 1], grid=False)
+        op_psiscale = self._spl_psi(o_points[:, 0], o_points[:, 1], grid=False)
         op_psiscale = 1 + (op_psiscale - np.min(op_psiscale)) / (np.max(op_psiscale) - np.min(op_psiscale))
 
         sortidx = np.argsort(op_dist * op_psiscale)
         # idx = np.argmin(op_dist)
         self._mg_axis = o_points[sortidx[0]]
-        self._psi_axis = np.asscalar(self.psi(R=self._mg_axis[0], Z=self._mg_axis[1]))
+        self._psi_axis = np.asscalar(self._spl_psi(self._mg_axis[0], self._mg_axis[1]))
         self._opoints = o_points[sortidx]
 
         # identify THE x-point as the x-point nearest in psi value to mg_axis
@@ -412,7 +409,7 @@ class Equilibrium(object):
         for i in np.arange(x_points.shape[0]):
             rxp = x_points[i, 0]
             zxp = x_points[i, 1]
-            psi_xp = np.asscalar(self.psi(R=rxp, Z=zxp, grid=False))
+            psi_xp = np.asscalar(self._spl_psi(rxp, zxp))
             psi_diff[i] = np.abs(psi_xp - self._psi_axis)
 
         xp_dist = (x_points[:, 0] - self._mg_axis[0]) ** 2 + (x_points[:, 1] - self._mg_axis[1]) ** 2
@@ -423,11 +420,11 @@ class Equilibrium(object):
         sortidx = np.argsort(psi_diff * xp_dist)
 
         self._x_point = x_points[sortidx[0]]
-        self._psi_lcfs = np.asscalar(self.psi(R=self._x_point[0], Z=self._x_point[1]))
+        self._psi_lcfs = np.asscalar(self._spl_psi(self._x_point[0], self._x_point[1]))
         self._x_points = x_points[sortidx]
 
         self._x_point2 = x_points[sortidx[1]]
-        self._psi_xp2 = np.asscalar(self.psi(R=self._x_point2[0], Z=self._x_point2[1]))
+        self._psi_xp2 = np.asscalar(self._spl_psi(self._x_point2[0], self._x_point2[1]))
 
         # get lcfs, for now using matplotlib contour line
 
@@ -463,3 +460,235 @@ class Equilibrium(object):
 class FluxSurface:
     def __init__(self, ):
         pass
+
+
+class Coordinates(object):
+
+    def __init__(self, equilibrium: Equilibrium, *coordinates, coord_type=None, grid=False, **coords):
+        self._eq = equilibrium
+        self._valid_coordinates = {'R', 'Z', 'psi_n', 'psi', 'rho'}
+        self._valid_coordinates_1d = {('psi_n',), ('psi',), ('rho',)}
+        self._valid_coordinates_2d = {('R', 'Z')}
+        self.dim = -1  # init only
+        self.grid = grid
+
+        self.__evaluate_input__(*coordinates, coord_type=coord_type, **coords)
+
+    def __call__(self, *args, **kwargs):
+        pass
+
+    def __iter__(self):
+        pass
+
+    def sort(self, order):
+        pass
+
+    @property
+    def R(self):
+        if self.dim == 2:
+            return self.x1
+
+    @property
+    def Z(self):
+        if self.dim == 2:
+            # todo
+            return self.x2
+
+    @property
+    def psi(self):
+        if self.dim == 1:
+            return self._eq._psi_axis + self.x1 * (self._eq._psi_lcfs - self._eq._psi_axis)
+        elif self.dim == 2:
+            return self._eq._spl_psi(self.x1, self.x2, grid=self.grid)
+
+    @property
+    def psi_n(self):
+        if self.dim == 1:
+            return self.x1
+        elif self.dim == 2:
+            return (self.psi - self._eq._psi_axis) / (self._eq._psi_lcfs - self._eq._psi_axis)
+
+    @property
+    def rho(self):
+        return np.sqrt(self.psi_n)
+
+    # todo
+    # @property
+    # def r_mid(self):
+    #     """
+    #     Midplane coordinate.
+    #     :return:
+    #     """
+    #
+    #
+    #     return
+
+    def getAs(self, coord_type=None):
+        if self.dim == 0:
+            return np.array(())
+        coord_type_ = self._verify_coord_type(coord_type)
+
+        if self.dim < len(coord_type):
+            raise ValueError('Can not return 2d from 1d data.')
+        raise ValueError('Not implemented yet.')
+
+    def __evaluate_input__(self, *coordinates, coord_type=None, **coords):
+        from collections import Iterable
+
+        if len(coordinates) == 0:
+            # todo:
+            self.dim = 0
+            xy = []
+            xy_name = []
+            for key, val in coords.items():
+                if key in self._valid_coordinates and val is not None:
+                    if isinstance(val, Iterable):
+                        if not isinstance(val, np.ndarray):
+                            val = np.array(val, ndmin=1)
+                        if len(val.shape) == 0:
+                            val = val.reshape((len(val), 1))
+                    else:
+                        val = np.array(val, ndmin=1)
+                    xy.append(val)
+                    xy_name.append(key)
+                    self.dim += 1
+
+            if self.dim == 0:
+                coord_type_ = ()
+            elif self.dim == 1:
+                self._x1_input = xy[0]
+                coord_type_ = tuple(xy_name)
+            elif self.dim == 2:
+                if tuple(xy_name) in self._valid_coordinates_2d:
+                    self._x1_input = xy[0]
+                    self._x2_input = xy[1]
+                    coord_type_ = tuple(xy_name)
+                elif tuple(xy_name[::-1]) in self._valid_coordinates_2d:
+                    self._x1_input = xy[1]
+                    self._x2_input = xy[0]
+                    coord_type_ = tuple(xy_name[::-1])
+                    if len(self._x1_input) != len(self._x2_input) and not self.grid:
+                        raise ValueError('All coordinates should contain same dimension.')
+                else:
+                    raise ValueError('Invalid combination of input coordinates.')
+
+            else:
+                # self._incompatible_dimension_error(self.dim)
+                raise ValueError('Operation in {} space has not be en allowed yet. Sorry.'
+                                 .format(self.dim))
+
+            self._coord_type_input = coord_type_
+        else:
+            if len(coordinates) == 1:
+                xy = coordinates[0]
+
+                if self.grid:
+                    print('WARNING: grid == True is not allowed for this coordinates input. '
+                          'Turning grid = False.')
+                self.grid = False
+
+                if isinstance(xy, Coordinates):
+                    # todo: ask Ondra (!)
+                    print('ERROR: not implemented yet')
+                    raise ValueError('not implemented yet')
+                    pass
+                elif isinstance(xy, Iterable):
+                    # input as array of size (N, dim),
+                    # if (N) add dimension
+                    if not isinstance(xy, np.ndarray):
+                        xy = np.array(xy, ndmin=2)
+                    if len(xy.shape) == 1:
+                        xy = xy.reshape((len(xy), 1))
+
+                    self.dim = xy.shape[1]
+                    if self.dim == 1:
+                        self._x1_input = xy[:, 0]
+                    elif self.dim == 2:
+                        self._x1_input = xy[:, 0]
+                        self._x2_input = xy[:, 1]
+                    else:
+                        self._incompatible_dimension_error(self.dim)
+                else:
+                    # 1d, one number
+                    self.dim = 1
+                    self._x1_input = np.array([xy])
+                pass
+            elif len(coordinates) == 2:
+                self.dim = 2
+                x1 = coordinates[0]
+                x2 = coordinates[1]
+
+                # assume _x1_input and _x2_input to be arrays of size (N)
+                if not isinstance(x1, np.ndarray):
+                    x1 = np.array(x1, ndmin=1)
+                if not isinstance(x2, np.ndarray):
+                    x2 = np.array(x2, ndmin=1)
+                self._x1_input = x1
+                self._x2_input = x2
+
+            else:
+                self._incompatible_dimension_error(len(coordinates))
+
+            self._coord_type_input = self._verify_coord_type(coord_type)
+        self._convert_to_default_coord_type()
+
+    def _verify_coord_type(self, coord_type):
+        if isinstance(coord_type, str):
+            coord_type = (coord_type,)
+
+        if self.dim == 0:
+            ret_coord_type = ()
+        elif self.dim == 1:
+            if coord_type is None:
+                ret_coord_type = ('psi_n',)
+            elif tuple(coord_type) in self._valid_coordinates_1d:
+                ret_coord_type = tuple(coord_type)
+            else:
+                ret_coord_type = ('psi_n',)
+                print("WARNING: _coord_type_input is not correct. \n"
+                      "{} is not allowed \n"
+                      "Force set _coord_type_input = ('psi_n',)"
+                      .format(tuple(coord_type)))
+        elif self.dim == 2:
+            if coord_type is None:
+                ret_coord_type = ('R', 'Z')
+            elif tuple(coord_type) in self._valid_coordinates_2d:
+                ret_coord_type = tuple(coord_type)
+            elif tuple(coord_type[::-1]) in self._valid_coordinates_2d:
+                ret_coord_type = tuple(coord_type[::-1])
+            else:
+                ret_coord_type = ('R', 'Z')
+                print("WARNING: _coord_type_input is not correct. \n"
+                      "{} is not allowed \n"
+                      "Force set _coord_type_input = ('R', 'Z')"
+                      .format(tuple(coord_type)))
+        else:
+            raise ValueError('Operation in {} space has not be en allowed yet. Sorry.'
+                             .format(self.dim))
+        # todo: make order dependent!
+        return ret_coord_type
+
+    def _incompatible_dimension_error(self, dim):
+        raise ValueError('Operation in {} space has not be en allowed yet. Sorry.'
+                         .format(dim))
+
+    def _convert_to_default_coord_type(self):
+        if self.dim == 0:
+            return
+        elif self.dim == 1:
+            # convert to psi_n
+            if self._coord_type_input == ('psi_n',):
+                self.x1 = self._x1_input
+            elif self._coord_type_input == ('psi',):
+                psi = self._x1_input
+                self.x1 = (psi - self._eq._psi_axis) / \
+                          (self._eq._psi_lcfs - self._eq._psi_axis)
+
+            elif self._coord_type_input == ('rho',):
+                self.x1 = self._x1_input ** 2
+            else:
+                raise ValueError('This should not happen.')
+        elif self.dim == 2:
+            # only (R, Z) coordinates are implemented now
+            self.x1 = self._x1_input
+            self.x2 = self._x2_input
